Renaming files and created new env for prod
