+++ api/controllers/admin_applications.py
